+++ src/lib/tvgShapeImpl.h
@@ -22,8 +22,8 @@
 #ifndef _TVG_SHAPE_IMPL_H_
 #define _TVG_SHAPE_IMPL_H_
 
+#include <memory.h>
 #include "tvgPaint.h"
-#include "tvgShapePath.h"
 
 /************************************************************************/
 /* Internal Class Implementation                                        */
@@ -54,6 +54,138 @@ struct ShapeStroke
     ~ShapeStroke()
     {
         if (dashPattern) free(dashPattern);
+    }
+};
+
+
+struct ShapePath
+{
+    PathCommand* cmds = nullptr;
+    uint32_t cmdCnt = 0;
+    uint32_t reservedCmdCnt = 0;
+
+    Point *pts = nullptr;
+    uint32_t ptsCnt = 0;
+    uint32_t reservedPtsCnt = 0;
+
+    ~ShapePath()
+    {
+        if (cmds) free(cmds);
+        if (pts) free(pts);
+    }
+
+    ShapePath() {}
+
+    ShapePath(const ShapePath* src)
+    {
+        cmdCnt = src->cmdCnt;
+        reservedCmdCnt = src->reservedCmdCnt;
+        ptsCnt = src->ptsCnt;
+        reservedPtsCnt = src->reservedPtsCnt;
+
+        cmds = static_cast<PathCommand*>(malloc(sizeof(PathCommand) * reservedCmdCnt));
+        if (!cmds) return;
+        memcpy(cmds, src->cmds, sizeof(PathCommand) * cmdCnt);
+
+        pts = static_cast<Point*>(malloc(sizeof(Point) * reservedPtsCnt));
+        if (!pts) {
+            free(cmds);
+            return;
+        }
+        memcpy(pts, src->pts, sizeof(Point) * ptsCnt);
+    }
+
+    void reserveCmd(uint32_t cmdCnt)
+    {
+        if (cmdCnt <= reservedCmdCnt) return;
+        reservedCmdCnt = cmdCnt;
+        cmds = static_cast<PathCommand*>(realloc(cmds, sizeof(PathCommand) * reservedCmdCnt));
+    }
+
+    void reservePts(uint32_t ptsCnt)
+    {
+        if (ptsCnt <= reservedPtsCnt) return;
+        reservedPtsCnt = ptsCnt;
+        pts = static_cast<Point*>(realloc(pts, sizeof(Point) * reservedPtsCnt));
+    }
+
+    void grow(uint32_t cmdCnt, uint32_t ptsCnt)
+    {
+        reserveCmd(this->cmdCnt + cmdCnt);
+        reservePts(this->ptsCnt + ptsCnt);
+    }
+
+    void reset()
+    {
+        cmdCnt = 0;
+        ptsCnt = 0;
+    }
+
+    void append(const PathCommand* cmds, uint32_t cmdCnt, const Point* pts, uint32_t ptsCnt)
+    {
+        memcpy(this->cmds + this->cmdCnt, cmds, sizeof(PathCommand) * cmdCnt);
+        memcpy(this->pts + this->ptsCnt, pts, sizeof(Point) * ptsCnt);
+        this->cmdCnt += cmdCnt;
+        this->ptsCnt += ptsCnt;
+    }
+
+    void moveTo(float x, float y)
+    {
+        if (cmdCnt + 1 > reservedCmdCnt) reserveCmd((cmdCnt + 1) * 2);
+        if (ptsCnt + 2 > reservedPtsCnt) reservePts((ptsCnt + 2) * 2);
+
+        cmds[cmdCnt++] = PathCommand::MoveTo;
+        pts[ptsCnt++] = {x, y};
+    }
+
+    void lineTo(float x, float y)
+    {
+        if (cmdCnt + 1 > reservedCmdCnt) reserveCmd((cmdCnt + 1) * 2);
+        if (ptsCnt + 2 > reservedPtsCnt) reservePts((ptsCnt + 2) * 2);
+
+        cmds[cmdCnt++] = PathCommand::LineTo;
+        pts[ptsCnt++] = {x, y};
+    }
+
+    void cubicTo(float cx1, float cy1, float cx2, float cy2, float x, float y)
+    {
+        if (cmdCnt + 1 > reservedCmdCnt) reserveCmd((cmdCnt + 1) * 2);
+        if (ptsCnt + 3 > reservedPtsCnt) reservePts((ptsCnt + 3) * 2);
+
+        cmds[cmdCnt++] = PathCommand::CubicTo;
+        pts[ptsCnt++] = {cx1, cy1};
+        pts[ptsCnt++] = {cx2, cy2};
+        pts[ptsCnt++] = {x, y};
+    }
+
+    void close()
+    {
+        if (cmdCnt > 0 && cmds[cmdCnt - 1] == PathCommand::Close) return;
+
+        if (cmdCnt + 1 > reservedCmdCnt) reserveCmd((cmdCnt + 1) * 2);
+        cmds[cmdCnt++] = PathCommand::Close;
+    }
+
+    bool bounds(float* x, float* y, float* w, float* h)
+    {
+        if (ptsCnt == 0) return false;
+
+        Point min = { pts[0].x, pts[0].y };
+        Point max = { pts[0].x, pts[0].y };
+
+        for(uint32_t i = 1; i < ptsCnt; ++i) {
+            if (pts[i].x < min.x) min.x = pts[i].x;
+            if (pts[i].y < min.y) min.y = pts[i].y;
+            if (pts[i].x > max.x) max.x = pts[i].x;
+            if (pts[i].y > max.y) max.y = pts[i].y;
+        }
+
+        if (x) *x = min.x;
+        if (y) *y = min.y;
+        if (w) *w = max.x - min.x;
+        if (h) *h = max.y - min.y;
+
+        return true;
     }
 };
